refactor: move Navbar to components
